enh: Add IDM version to template (#44)
## Description
This introduces the new IDM versioning system.
After this is merged I will release chef-cli 0.8.0, which will support
this format but will assume v0.1.0 if no version is provided. Template
repo should default the users to the latest version.

## DevRev issue
https://app.devrev.ai/devrev/works/ISS-188976

## Documentation PR
I'll document this in the GitHub release as well as announcement,
documentation PR will come later.