Update make scripts (#48)
# Description
This pull request refactors the `deploy` and `cleanup` scripts for
managing snap-in packages and versions. The main change is the removal
of sourcing `shared.sh` file, which is no longer in use, and the
`$DR_OPTS` variable, which was previously used to pass additional
options to the `devrev` CLI commands.

# DevRev issue
https://app.devrev.ai/devrev/works/ISS-203111

# Documentation PR
https://github.com/devrev/fern-api-docs/pull/293

Preview:
https://devrev-preview-b36b1ce8-1f2b-4b9e-b71b-00e4a027cfc0.docs.buildwithfern.com/airdrop/deploy-to-organization